MDL-51723 Enroll: LDAP default search filter is wrongly used.
The search filter is correctly configured in the constructor but
when passed to ldaplib.php::ldap_find_userdn() it was taken straight from
ldap_getdefaults(), ignoring what optionally configured as a custom setting
and how complex could be its setup, already well done in the constructor.
When MDL-47004 changed the default value for 'ad' using an equality filter
the enrollment against an Active Directory source stopped working.